code_examples/python_libarary/requirements.txt
opencv-python
matplotlib
numpy
albumentations
pyyaml
torch
torchvision